Refactor pull_issues into a new sync module
