+++ test/conftest.py
@@ -16,14 +16,14 @@ def mock_jira():
     - Mock the config object from writing to disk
     - Mock the pypi jira module's Jira object. Consuming tests can mock individual methods as
       necessary
-    - Mock the _connect method
+    - Mock the connect method
     - Mock the write_issues method
     """
     jira = Jira()
     jira.config = AppConfig(username='test', password='dummy', projects={'CNTS'})
     jira.config.write_to_disk = mock.Mock()
-    jira._connect = mock.Mock()
     jira._jira = mock.Mock(spec=mod_jira.JIRA)
+    jira.connect = mock.Mock(return_value=jira._jira)
     jira.write_issues = mock.Mock()
     jira.load_issues = mock.Mock()
     return jira
